feat: created edit album view
