+++ musicApp/templates/albums/album-details.html
@@ -5,26 +5,24 @@
 <div class="wrapper">
   <div class="albumCover">
     <!-- Album Image -->
-    <img src="/images/Lorde.jpg" alt="Cover Image" />
+    <img src="{{ album.image_url }}" alt="Cover Image" />
   </div>
 
   <div class="albumInfo">
     <div class="albumText">
       <!-- Album Info -->
-      <h1>Name: Melodrama</h1>
-      <h3>Artist: Lorde</h3>
-      <h4>Genre: Pop Music</h4>
-      <h4>Price: $7.33</h4>
+      <h1>Name: {{ album.album_name }}</h1>
+      <h3>Artist: {{ album.artist }}</h3>
+      <h4>Genre: {{ album.genre }}</h4>
+      <h4>Price: ${{ album.price|floatformat:2 }}</h4>
       <p>
-        Description: Melodrama is the second studio album by New Zealand
-        singer-songwriter Lorde. It was released on 16 June 2017 by Lava
-        and Republic Records and distributed through Universal.
+        Description: {{ album.description }}
       </p>
     </div>
 
     <div class="actionBtn">
       <!-- Album Buttons -->
-      <a href="#" class="edit">Edit</a>
+      <a href="{% url 'albums:edit' album.pk %}" class="edit">Edit</a>
       <a href="#" class="remove">Delete</a>
     </div>
   </div>
